Separate DB Sync functions
