+++ functions/src/modules/SQR.functions.ts
@@ -6,19 +6,6 @@ const db = admin.database();
 import * as helpers from './../helpers';
 
 /////////////////////////////////////////////////
-//
-//   Add MP3 name to DB (Storage Upload Trigger)
-//
-/////////////////////////////////////////////////
-
-export const importFilesFromStorage = functions.storage.object()
-.onFinalize( object => {
-    const filePath = object.name;
-    helpers.storeFileToDB(filePath, db, 'soundQualityReporting');
-    return 1;
-});
-
-/////////////////////////////////////////////////
 //          OnNewAllotment (DB create and update Trigger)
 //      1. Mark the files in the database --> { status: "Given" }
 //              Function --> UpdateFilesOnNewAllotment
@@ -91,45 +78,6 @@ export const sendEmailOnNewAllotment = functions.database.ref('/sqr/allotments/{
         }
     
     return 1;
-});
-
-
-/////////////////////////////////////////////////
-//          Sync Storage to DB (HTTP Trigger)
-//
-//      1. Add the currently uploaded MP3s into the DB (handleCurrentlyUploadedFiles)
-//
-//      2. Remove DB entries for MP3s that don't exist (removeNonExistingMp3DBEntries)
-/////////////////////////////////////////////////
-export const syncStorageToDB = functions.https.onRequest( async (req, res) => {
-    ///////////////////////////////////////////////////////
-    //      1. Add the currently uploaded MP3s into the DB
-    ///////////////////////////////////////////////////////
-    const bucketFiles = await bucket.getFiles();
-    bucketFiles.forEach(innerFilesObject => {
-        innerFilesObject.forEach(file => {
-            helpers.storeFileToDB(file.name, db, 'soundQualityReporting');
-        });
-    });
-
-    ///////////////////////////////////////////////////////
-    //      2. Remove DB entries for MP3s that don't exist
-    ///////////////////////////////////////////////////////
-    const filesSnapshot = await db.ref(`/files`).once("value");
-    let files = filesSnapshot.val();
-
-    for (let list in files) {
-        for (let file in files[list]) {
-            const existingBucketFiles = await bucket.file(`/mp3/${list}/${file}.mp3`).exists();
-            // **Found** in DB but not in STORAGE
-            // Removing should be done only if the `status` is `Spare`
-            if (!existingBucketFiles[0] && files[list][file]['soundQualityReporting'].status === 'Spare') 
-                helpers.removeFromDB(db, `/files/${list}/${file}`)
-
-        }
-    }
-
-    return res.send(`Started Execution, the process is now Running in the background`);
 });
